remove dead code from metering API handlers
